AI ASSIST/requirements.txt
# BFSI Call Center AI Assistant - Dependencies
# Python 3.9+ recommended

# Core ML
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
accelerate>=0.24.0

# RAG and retrieval
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Fine-tuning
datasets>=2.14.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Utilities
pyyaml>=6.0
numpy>=1.24.0
